Improving PATH detection on Windows
To get the shell execution to work properly on windows we need Python to
execute commands with the shell variable. This is slightly less secure,
but I'm not aware of another way to get the Windows PATH var right now.
This allows us to execute zabbix_sender regardless of where it is
installed.