Prevent Chrome security messages
We're running several flags through Chromedriver that make Chrome
worried. This worry causes the browser to put up a message from time to
time about stability and security. This test-type flag is reported to
tell chrome to chill out and stop warning us. This is ideal for the
museum environment where warnings are not idea for our visitors.