Adding an uptime startup delay
This adds in a configurable delay, which is implemented when the
computer boots. This solves a problem, where the browser would try to
load the udnerlying server process right when the computer boots. It
wouldn't be ready, which would cause the browser to load with a white
screen.

We can't delay the launch process in upstart(linux) or launchd(OSX) and
still use the keep alive functions. So we handle the delay in here.