Removing the node modules folder
This actually presented problems in building the application for
multiple OS environments. If we include the node_modules folder we can't
build for linux and Mac. The npm install process works on both of those
environments, so we'll go with working but imprefect over future proof
and not working now.