Remove testing framework and hello world
Since we're loading our content from a separate application via the
config file, we don't need the app file for displaying the Hello World
copy.

The end to end tests expect this Hello World setup. We're not really
using a testing framework across our projects right now, so we're also
removing this system. We'll examine options in #6

This continues work on #6.