feat: Add customized application icons
Picking a hand because it's easy to find, distinctive, and works for
now.

I colored the default icon in a little bit and described the original
source in the README

I'm also removing a bunch of the old app icon types that aren't used by
the electron build system any longer. We only need these few files.

Issue: #4