feat: Implment react in our webpack setup
Update our webpack config to get react code to compile into the dist
folder along with the Electron app. Renaming the files to help
differentiate the purposes.

Moving React to dependencies from devDep. We're using it in the
compiled app so it's a requirement for the production environment.

Update the main location, to point to the electron build file.

Move babel configration into the webpack setup, where it's used.

Update some packages and remove some others that aren't in use
in our current webpack system.

This is work in progrsss on implementing settings page:
Issue #3