fix: Reenable shortcuts win & linux kiosk modes
Handle shortcuts using globalShortcut on Windows and Linux kiosks. The
local shortcuts are tied to the menu system, so we can't use those.
There are local shortcut plugins, but they don't seem to be working
properly in my test Windows environment. We're going to use global
shortcuts for now. It's not ideal, but it's not a big deal in the
production environment.

Issues: #46