feat: Validate URL in settings form
Check the URL, requiring the protocol either http or https on the
settings page.
Update the warning message to help the user handle the error.

This will help people setup the app correctly. Seems like it's better to
require people to enter full URLs instead of just example.org and then
have to handle all the edge cases where this may or may not be valid.

Issues: #22