fix: Improve window handling
Store all of the display information in a single object that matches the
display information. Don't throw any of the display information away,
just add URLs to it. This cleans up the startup process and fixes an
issue where the app couldn't find a proper setting when it first starts
up without any settings.

Issues: #53