feat: Add warning about disconnected displays
When a disconnected display is also enabled, we should alert the user to
how the app will behave. It will stay enabled, but it won't be shown
unless the display that it is linked to is connected. Adding a tooltop
to make this clear.

Trygve and I talked about this in person and agree that for most of the
use cases for this we'd rather have the app behave this way.

Issues: #68