feat: Improve settings error handling
We want the error message for missing URL to appear as soon as a display
is enabled. Don't wait until the form element is touched to show the
error. This makes the app a little clearer to use and also displays an
error if you try to submit the form without the right URL displayed,
before you've touched the URL field.