feat: Upgrading Electron to address node crashes
We're seeing some crashes related to the serial port module. Upgrading
Electron seems to address them. I believe that this is related to the
version of node that electron is running against. It's also good to get
up to speed with the latest Electron for other improvements.

The upgrade doesn't cause any problems with any existing functionality
in my testing.

Issues: #82, #47