feat: Improve Stele package naming structure
Adding a template to name the various OS files the same with only minor
changes based on the desired file type. This will ensure that end users
know how to install the app more consistently. We ran into some
situations where the win exe setup app was getting installed as the main
app, not being used once for install. This will help clarify that.

Issues: #78, #82