feat: Stop publishing Linux releases
We're stoping our support of Linux for now. We don't have any active
projects using Linux for kiosks. We don't have the resources to test
each build on Linux right now, so we shouldn't be publishing releases
that might not work.

Leaving some linux specific workarounds in the code because they don't
affect the other builds in any way.